Rename `StandardKit` to `StandardModel`
