+++ StandardModelTests/Chromodynamics/Collection+Pair.swift
