+++ StandardModelTests/Chromodynamics/Collection+PairTests.swift
