+++ StandardModelTests/Chromodynamics/Collection+Repetition.swift
