+++ StandardModelTests/SymmetryTests.swift
@@ -15,17 +15,37 @@
 // not, see https://www.gnu.org/licenses.
 // ===-------------------------------------------------------------------------------------------===
 
+import Foundation
+import Numerics
 import Testing
 
-@testable import StandardKit
+@testable import StandardModel
 
-struct AnticolorTests {
-  @Test
-  func redIsCounterpartOfAntired() { #expect(Anti(red).counterpart === red) }
+struct SymmetryTests {
+  @Suite("U1")
+  struct U1Tests {
+    @Test
+    func fieldIsUntransformedWhenUnrotated() {
+      #expect(Complex(2, 4).u1(by: .zero) == Complex(2, 4))
+    }
 
-  @Test
-  func greenIsCounterpartOfAntigreen() { #expect(Anti(green).counterpart === green) }
+    @Test(
+      arguments: stride(from: 2, to: 64, by: 2).map {
+        Measurement(value: .pi * $0, unit: UnitAngle.radians)
+      }
+    )
+    func fieldIsUntransformedUponFullTurn(of angle: Measurement<UnitAngle>) {
+      #expect(Complex(2, 4).u1(by: angle).isApproximatelyEqual(to: Complex(2, 4)))
+    }
 
-  @Test
-  func blueIsCounterpartOfAntiblue() { #expect(Anti(blue).counterpart === blue) }
+    @Test
+    func fieldIsTransformedWhenRotatedByNonGroupIdentity() {
+      #expect(
+        Complex(2, 4).u1(by: Measurement(value: 2, unit: UnitAngle.radians)).isApproximatelyEqual(
+          to: Complex(-4.46, 0.15),
+          relativeTolerance: 0.01
+        )
+      )
+    }
+  }
 }
